Refractor tests
