+++ hypha/apply/projects/reports/tests/utils.py
@@ -1,0 +1,13 @@
+# A boilerplate stream form for Project Report tests
+FORM_FIELDS = [
+    {
+        "id": "012a4f29-0882-4b1c-b567-aede1b601d4a",
+        "type": "number",
+        "value": {
+            "required": True,
+            "help_text": "",
+            "field_label": "How many folks did you reach?",
+            "default_value": "",
+        },
+    }
+]
