+++ hypha/apply/projects/tests/test_views.py
@@ -1,14 +1,11 @@
 import json
 from io import BytesIO
 
-import factory
-from dateutil.relativedelta import relativedelta
 from django.conf import settings
 from django.contrib.auth.models import AnonymousUser
 from django.core.exceptions import PermissionDenied
 from django.test import RequestFactory, TestCase, override_settings
 from django.urls import reverse
-from django.utils import timezone
 
 from hypha.apply.funds.tests.factories import LabSubmissionFactory
 from hypha.apply.projects.utils import get_invoice_status_display_value
@@ -25,19 +22,15 @@ from hypha.apply.users.tests.factories import (
 from hypha.apply.utils.testing.tests import BaseViewTestCase
 from hypha.home.factories import ApplySiteFactory
 
-from ...funds.models.forms import ApplicationBaseProjectReportForm
 from ..forms import SetPendingForm
 from ..models.payment import CHANGES_REQUESTED_BY_STAFF, DECLINED, SUBMITTED
 from ..models.project import (
     APPROVE,
-    CLOSING,
-    COMPLETE,
     CONTRACTING,
     DRAFT,
     INTERNAL_APPROVAL,
     INVOICING_AND_REPORTING,
     REQUEST_CHANGE,
-    ProjectReportForm,
     ProjectSettings,
 )
 from ..views.project import ContractsMixin, ProjectDetailApprovalView
@@ -49,9 +42,6 @@ from .factories import (
     PAFApprovalsFactory,
     PAFReviewerRoleFactory,
     ProjectFactory,
-    ReportConfigFactory,
-    ReportFactory,
-    ReportVersionFactory,
     SupportingDocumentFactory,
 )
 
@@ -1274,570 +1264,6 @@ class TestProjectListView(TestCase):
 
         response = self.client.get(url, follow=True)
         self.assertEqual(response.status_code, 403)
-
-
-class TestStaffSubmitReport(BaseViewTestCase):
-    base_view_name = "edit"
-    url_name = "funds:projects:reports:{}"
-    user_factory = StaffFactory
-    report_form_id = None
-
-    def setUp(self):
-        super().setUp()
-        report_form, _ = ProjectReportForm.objects.get_or_create(
-            name=factory.Faker("word"),
-            form_fields=FORM_FIELDS,
-        )
-        self.report_form_id = report_form.id
-
-    def get_kwargs(self, instance):
-        return {
-            "pk": instance.pk,
-        }
-
-    def test_get_page_for_inprogress_project(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.get_page(report)
-        self.assertContains(response, report.project.title)
-
-    def test_cant_get_page_for_closing_and_complete_project(self):
-        report = ReportFactory(project__status=CLOSING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-        report = ReportFactory(project__status=COMPLETE)
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_submit_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "11"}
-        )
-        report.refresh_from_db()
-        self.assertRedirects(response, report.project.get_absolute_url())
-        self.assertEqual(
-            report.versions.first().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "11"},
-        )
-        self.assertEqual(report.versions.first(), report.current)
-        self.assertEqual(report.current.author, self.user)
-        self.assertIsNone(report.draft)
-
-    def test_cant_submit_report_for_closing_and_complete_project(self):
-        report = ReportFactory(project__status=CLOSING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "13"}
-        )
-        self.assertEqual(response.status_code, 403)
-
-        report = ReportFactory(project__status=COMPLETE)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "13"}
-        )
-        self.assertEqual(response.status_code, 403)
-
-    def test_submit_private_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING)
-        # Link the single-field report_form to the Fund associated with this Submission.
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "17"}
-        )
-        report.refresh_from_db()
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            report.versions.first().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "17"},
-        )
-        self.assertEqual(report.versions.first(), report.current)
-        self.assertEqual(report.current.author, self.user)
-        self.assertIsNone(report.draft)
-
-    def test_cant_submit_blank_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(report, {})
-        report.refresh_from_db()
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(report.versions.count(), 0)
-
-    def test_save_report_draft(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "19", "save": "Save"}
-        )
-        report.refresh_from_db()
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            report.versions.first().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "19"},
-        )
-        self.assertEqual(report.versions.first(), report.draft)
-        self.assertIsNone(report.current)
-
-    def test_save_report_with_draft(self):
-        report = ReportFactory(is_draft=True, project__status=INVOICING_AND_REPORTING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        self.assertEqual(report.versions.first(), report.draft)
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "23"}
-        )
-        report.refresh_from_db()
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            report.versions.last().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "23"},
-        )
-        self.assertEqual(report.versions.last(), report.current)
-        self.assertIsNone(report.draft)
-
-    def test_edit_submitted_report(self):
-        report = ReportFactory(
-            is_submitted=True,
-            project__status=INVOICING_AND_REPORTING,
-            form_fields=json.dumps(FORM_FIELDS),
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        self.assertEqual(report.versions.first(), report.current)
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "29", "save": " Save"}
-        )
-        report.refresh_from_db()
-        self.assertRedirects(response, report.project.get_absolute_url())
-        self.assertEqual(
-            report.versions.last().form_data["012a4f29-0882-4b1c-b567-aede1b601d4a"],
-            "29",
-        )
-        self.assertEqual(report.versions.last(), report.draft)
-        self.assertEqual(report.versions.first(), report.current)
-
-    def test_resubmit_submitted_report(self):
-        yesterday = timezone.now() - relativedelta(days=1)
-        version = ReportVersionFactory(
-            report__project__status=INVOICING_AND_REPORTING,
-            submitted=yesterday,
-        )
-        report = version.report
-        report.form_fields = json.dumps(FORM_FIELDS)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        report.current = version
-        report.submitted = version.submitted
-        report.save()
-        self.assertEqual(report.submitted, yesterday)
-        self.assertEqual(report.versions.first(), report.current)
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "31"}
-        )
-        report.refresh_from_db()
-        self.assertRedirects(response, report.project.get_absolute_url())
-        self.assertEqual(
-            report.versions.last().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "31"},
-        )
-        self.assertEqual(report.versions.last(), report.current)
-        self.assertIsNone(report.draft)
-        self.assertEqual(report.submitted.date(), yesterday.date())
-        self.assertEqual(report.current.submitted.date(), timezone.now().date())
-
-    def test_cant_submit_future_report(self):
-        submitted_report = ReportFactory(
-            end_date=timezone.now() + relativedelta(days=1),
-            is_submitted=True,
-            project__status=INVOICING_AND_REPORTING,
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=submitted_report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        future_report = ReportFactory(
-            end_date=timezone.now() + relativedelta(days=3),
-            project=submitted_report.project,
-        )
-        response = self.post_page(
-            future_report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "37"}
-        )
-        self.assertEqual(response.status_code, 403)
-
-
-class TestApplicantSubmitReport(BaseViewTestCase):
-    base_view_name = "edit"
-    url_name = "funds:projects:reports:{}"
-    user_factory = ApplicantFactory
-    report_form_id = None
-
-    def setUp(self):
-        super().setUp()
-        report_form, _ = ProjectReportForm.objects.get_or_create(
-            name=factory.Faker("word"),
-            form_fields=FORM_FIELDS,
-        )
-        self.report_form_id = report_form.id
-
-    def get_kwargs(self, instance):
-        return {
-            "pk": instance.pk,
-        }
-
-    def test_get_own_report_for_inprogress_project(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, project__user=self.user
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.get_page(report)
-        self.assertContains(response, report.project.title)
-
-    def test_cant_get_own_report_for_closing_and_complete_project(self):
-        report = ReportFactory(project__status=CLOSING, project__user=self.user)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-        report = ReportFactory(project__status=COMPLETE, project__user=self.user)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_cant_get_other_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_submit_own_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, project__user=self.user
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "37"}
-        )
-        report.refresh_from_db()
-        self.assertRedirects(response, report.project.get_absolute_url())
-        self.assertEqual(
-            report.versions.first().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "37"},
-        )
-        self.assertEqual(report.versions.first(), report.current)
-        self.assertEqual(report.current.author, self.user)
-
-    def test_submit_private_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, project__user=self.user
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "41"}
-        )
-        report.refresh_from_db()
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            report.versions.first().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "41"},
-        )
-        self.assertEqual(report.versions.first(), report.current)
-        self.assertEqual(report.current.author, self.user)
-        self.assertIsNone(report.draft)
-
-    def test_cant_submit_blank_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, project__user=self.user
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(report, {})
-        report.refresh_from_db()
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(report.versions.count(), 0)
-
-    def test_save_report_draft(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, project__user=self.user
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "43", "save": "Save"}
-        )
-        report.refresh_from_db()
-        self.assertRedirects(response, report.project.get_absolute_url())
-        self.assertEqual(
-            report.versions.first().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "43"},
-        )
-        self.assertEqual(report.versions.first(), report.draft)
-        self.assertIsNone(report.current)
-
-    def test_save_report_with_draft(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING,
-            is_draft=True,
-            project__user=self.user,
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        self.assertEqual(report.versions.first(), report.draft)
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "47"}
-        )
-        report.refresh_from_db()
-        self.assertRedirects(response, report.project.get_absolute_url())
-        self.assertEqual(
-            report.versions.last().form_data,
-            {"012a4f29-0882-4b1c-b567-aede1b601d4a": "47"},
-        )
-        self.assertEqual(report.versions.last(), report.current)
-        self.assertIsNone(report.draft)
-
-    def test_cant_edit_submitted_report(self):
-        # :todo: need to check if applicant can edit submitted report or not.
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING,
-            is_submitted=True,
-            project__user=self.user,
-        )
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        self.assertEqual(report.versions.first(), report.current)
-        response = self.post_page(
-            report, {"public_content": "Some text", "save": " Save"}
-        )
-        self.assertEqual(response.status_code, 403)
-
-    def test_cant_submit_other_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING)
-        ApplicationBaseProjectReportForm.objects.get_or_create(
-            application_id=report.project.submission.page.specific.id,
-            form_id=self.report_form_id,
-        )
-        response = self.post_page(
-            report, {"012a4f29-0882-4b1c-b567-aede1b601d4a": "53"}
-        )
-        self.assertEqual(response.status_code, 403)
-
-
-class TestStaffReportDetail(BaseViewTestCase):
-    base_view_name = "detail"
-    url_name = "funds:projects:reports:{}"
-    user_factory = StaffFactory
-
-    def get_kwargs(self, instance):
-        return {
-            "pk": instance.pk,
-        }
-
-    def test_can_access_submitted_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, is_submitted=True
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 200)
-
-        report = ReportFactory(project__status=CLOSING, is_submitted=True)
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 200)
-
-        report = ReportFactory(project__status=COMPLETE, is_submitted=True)
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 200)
-
-    def test_cant_access_skipped_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING, skipped=True)
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_cant_access_draft_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING, is_draft=True)
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_cant_access_future_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING,
-            end_date=timezone.now() + relativedelta(days=1),
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-
-class TestApplicantReportDetail(BaseViewTestCase):
-    base_view_name = "detail"
-    url_name = "funds:projects:reports:{}"
-    user_factory = StaffFactory
-
-    def get_kwargs(self, instance):
-        return {
-            "pk": instance.pk,
-        }
-
-    def test_can_access_own_submitted_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING,
-            is_submitted=True,
-            project__user=self.user,
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 200)
-
-    def test_cant_access_own_draft_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING,
-            is_draft=True,
-            project__user=self.user,
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_cant_access_own_future_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING,
-            end_date=timezone.now() + relativedelta(days=1),
-            project__user=self.user,
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_cant_access_other_submitted_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, is_submitted=True
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 200)
-
-    def test_cant_access_other_draft_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING, is_draft=True)
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-    def test_cant_access_other_future_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING,
-            end_date=timezone.now() + relativedelta(days=1),
-        )
-        response = self.get_page(report)
-        self.assertEqual(response.status_code, 403)
-
-
-class TestSkipReport(BaseViewTestCase):
-    base_view_name = "skip"
-    url_name = "funds:projects:reports:{}"
-    user_factory = StaffFactory
-
-    def get_kwargs(self, instance):
-        return {
-            "pk": instance.pk,
-        }
-
-    def test_can_skip_report(self):
-        report = ReportFactory(project__status=INVOICING_AND_REPORTING, past_due=True)
-        response = self.post_page(report)
-        self.assertEqual(response.status_code, 200)
-        report.refresh_from_db()
-        self.assertTrue(report.skipped)
-
-    def test_can_unskip_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, skipped=True, past_due=True
-        )
-        response = self.post_page(report)
-        self.assertEqual(response.status_code, 200)
-        report.refresh_from_db()
-        self.assertFalse(report.skipped)
-
-    def test_cant_skip_current_report(self):
-        config = ReportConfigFactory(disable_reporting=False)
-        report = ReportFactory(
-            project=config.project,
-            project__status=INVOICING_AND_REPORTING,
-            end_date=timezone.now() + relativedelta(days=1),
-        )
-        response = self.post_page(report)
-        self.assertEqual(response.status_code, 200)
-        report.refresh_from_db()
-        self.assertFalse(report.skipped)
-
-    def test_cant_skip_submitted_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, is_submitted=True
-        )
-        response = self.post_page(report, data={})
-        self.assertEqual(response.status_code, 200)
-        report.refresh_from_db()
-        self.assertFalse(report.skipped)
-
-    def test_can_skip_draft_report(self):
-        report = ReportFactory(
-            project__status=INVOICING_AND_REPORTING, is_draft=True, past_due=True
-        )
-        response = self.post_page(report)
-        self.assertEqual(response.status_code, 200)
-        report.refresh_from_db()
-        self.assertTrue(report.skipped)
 
 
 class TestStaffProjectDetailDownloadView(BaseViewTestCase):
